Move json classes into Json namespace
