+++ visitz/VisitzApi/Json/PayloadOptions.cs
@@ -1,0 +1,11 @@
+using System.Text.Json;
+
+namespace VisitzApi.Json;
+
+public static class PayloadOptions
+{
+	public static readonly JsonSerializerOptions Default = new()
+	{
+		PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
+	};
+}
