--- visitz/VisitzApi/JsonKey.cs
@@ -1,12 +1,0 @@
-﻿namespace VisitzApi
-{
-    internal class JsonKey
-    {
-        public static readonly string DocRequest = "docRequest";
-        public static readonly string PayLoad = "payLoad";
-        public static readonly string Payload = "payload";
-        public static readonly string Status = "status";
-        public static readonly string StatusResponse = "statusResponse";
-        public static readonly string Success = "Success";
-    }
-}
