+++ visitz/VisitzApi/Requests/GetCaseloadEndpoint.cs
@@ -1,11 +1,12 @@
 ﻿using System.Text.Json;
 using System.Text.Json.Nodes;
 using VisitzApi.ErrorHandling;
+using VisitzApi.Json;
 using VisitzApi.Models;
 
 namespace VisitzApi.Requests
 {
-    internal class GetCaseloadEndpoint(string baseUrl, params string[] workerIds) 
+	internal class GetCaseloadEndpoint(string baseUrl, params string[] workerIds) 
         : VisitzBaseEndpoint<IEnumerable<CaseloadEntity>>(baseUrl, CaseloadPath)
     {
         private static readonly string CaseloadPath = "/v1/620b";
