+++ visitz/VisitzApi/Requests/SubmitSafetyAssessmentEndpoint.cs
@@ -1,5 +1,6 @@
 ﻿using System.Text.Json;
 using System.Text.Json.Nodes;
+using VisitzApi.Json;
 using VisitzApi.Models.SafetyAssess;
 
 namespace VisitzApi.Requests;
